Rewrite the build/deploy process
The old build and deploy process committed the built artefacts into Git.
We do not want to continue doing so. The new process creates the app as
`.tgz` file containing the `dist` outputs, and copies that to S3.

Also took a pass over the README to bring it up to date.